Bug fix  new features restructuring
There was a bug in the Eikonal solver in the square root case - fixed.
The program knows the concept of people density and velocity
Mapping from density to velocity is implemented
Several test cases added, that address the relation of density and velocity and use the iconal solver on areas of reduced velocity.